fix: split github_app_auth_test.go across handlers and feedback packages
The extraction PR #18095 moved classification tests to the feedback
sub-package, but they reference unexported types (searchApp, searchItem,
classifyIssue) still in handlers. Split the test: token/slug tests stay
in feedback, classification tests return to handlers with qualified
feedback.DefaultConsoleAppSlug references. Add RoundTripFunc test helper
to feedback package.
