+++ pkg/api/handlers/feedback/testenv_test.go
@@ -1,0 +1,9 @@
+package feedback
+
+import "net/http"
+
+type RoundTripFunc func(req *http.Request) *http.Response
+
+func (f RoundTripFunc) RoundTrip(req *http.Request) (*http.Response, error) {
+	return f(req), nil
+}
